+++ src/sw/allotmentclub/alembic/versions/add_deleted_flag_to_external_recipients_7e15fdd9747a.py
@@ -1,0 +1,24 @@
+"""Add deleted flag to external recipients
+
+Revision ID: 7e15fdd9747a
+Revises: 2952c5487114
+Create Date: 2021-01-11 07:01:46.325174
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+# revision identifiers, used by Alembic.
+revision = '7e15fdd9747a'
+down_revision = '2952c5487114'
+
+
+def upgrade():
+    op.add_column(
+        'externalrecipient',
+        sa.Column('deleted', sa.Boolean(), nullable=True)
+    )
+
+
+def downgrade():
+    op.drop_column('externalrecipient', 'deleted')