+++ src/sw/allotmentclub/browser/external.py
@@ -4,8 +4,33 @@ from ..log import user_data_log, log_with_user
 from .base import to_string
 from pyramid.view import view_config
 from sw.allotmentclub import ExternalRecipient
+from sqlalchemy import or_, and_
 import collections
 import sw.allotmentclub.browser.base
+
+
+def filter_external_recipients(query):
+    return (
+        query.filter(
+            or_(
+                and_(
+                    ExternalRecipient.city != '',
+                    ExternalRecipient.city.isnot(None),
+                ),
+                and_(
+                    ExternalRecipient.email != '',
+                    ExternalRecipient.email.isnot(None),
+                )
+            )
+        )
+        .filter(
+            and_(
+                ExternalRecipient.lastname != '',
+                ExternalRecipient.lastname.isnot(None),
+            )
+        )
+        .filter(ExternalRecipient.deleted.isnot(True))
+    )
 
 
 class Query(sw.allotmentclub.browser.base.Query):
@@ -21,7 +46,7 @@ class Query(sw.allotmentclub.browser.base.Query):
     }
 
     def select(self):
-        return (
+        return filter_external_recipients(
             self.db.query(
                 ExternalRecipient.id.label('#'),
                 ExternalRecipient.organization.label('Firma'),
@@ -48,7 +73,9 @@ class ExternalRecipientListView(sw.allotmentclub.browser.base.TableView):
         dict(url='external_add', btn_class='btn-success', icon='fa fa-plus',
              title='Neu'),
         dict(url='external_edit', btn_class='btn-success', icon='fa fa-pencil',
-             title='Bearbeiten')]
+             title='Bearbeiten'),
+        dict(url='external_delete', btn_class='btn-danger',
+             icon='glyphicon glyphicon-trash', title='Löschen')]
 
 
 @view_config(route_name='external_edit', renderer='json',
@@ -124,3 +151,11 @@ class ExternalRecipientAddView(ExternalRecipientEditView):
     @property
     def route_name(self):
         return 'external_edit'
+
+
+@view_config(route_name='external_delete', renderer='json',
+             permission='view')
+class ExternalRecipientDeleteView(sw.allotmentclub.browser.base.DeleteView):
+
+    model = ExternalRecipient
+    mark_deleted = True